Cleanup.
- Breadboard into more files (that class is getting very big, but for now, I don't see how splitting everything up would be better).
- Fix warnings (except for lib/protocol)
- device/configurations: renamed to types.h, only the structs that are actually used outside of device remain. Everything else moved into respective class (device/oled)
- breadboard/types.h: structs moved to breadboard, as only internally used.
